requirements.txt
boto3
requests
beautifulsoup4
botocore
aws-lambda-powertools
aws-xray-sdk
pysuitecrm
oauthlib
moto
pytest-mock
cryptography==42.0.0
httpx==0.27.0
